pyproject.toml
[tool.poetry]
name = "signalhooks"
version = "0.1.4"
description = "Set of useful hooks that you can attach to your Django Signals to notify other services when a Signal is triggered"
authors = ["Martin Zugnoni <martin.zugnoni@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7, <4"
Django = ">=2.2"
boto3 = ">=1.14"
requests = ">=2.23"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
ipdb = "^0.13.2"
moto = "^1.3.14"
pytest = "^5.4.3"
pytest-django = "^3.9.0"
responses = "^0.10.15"
coverage = ">=5.1"
pytest-cov = "^2.10.0"
pylint = "^2.5.3"
codecov = "^2.1.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
